REFRACTOR: few files refractoring and structured
